Moved MoedaCursorAdapter
